Update prost, toolchain, and adjust clippy action (#128)
* Update prost, toolchain, and adjust clippy action

* fix clippy warnings

+++ src/stream/source/socket.rs
@@ -216,7 +216,7 @@ mod tests {
 
         wait(1);
         sink_comp.on_definition(|cd| {
-            assert_eq!(cd.data.len(), (1 as usize));
+            assert_eq!(cd.data.len(), (1_usize));
             assert_eq!(cd.watermarks.last().unwrap().timestamp, 1);
         });
     }
